pyproject.toml
[tool.poetry]
name = "andromeda-desktop"
version = "0.0.1"
description = ""
authors = ["lucas <medeirosrosalucas@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.9"
python-decouple = "3.4"
astropy = "*"
Pillow = "8.3.2"

[tool.poetry.dev-dependencies]
pytest = "6.2.4"
pytest-asyncio = "0.15.1"
pytest-mock = "3.6.1"
coverage = "4.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"